OrderStatus Crud

+++ src/main/java/com/luxrest/rm/OrderStatus/OrderStatusRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderStatusRepository extends JpaRepository<OrderStatus, Long> {
+public interface OrderStatusRepository extends JpaRepository<OrderStatus, Integer> {
 }
